Refactor: Introduce getDecimalsData helper function and integrate it into automationV2Tvl for improved token handling
